finalizado curso de angular criando crud de redes sociais e criado frontend.

+++ BackEnd/EventPro.Api/Helpers/Utils.cs
@@ -15,9 +15,12 @@ public class Utils : IUtils
 
     public void DeleteImage(string imageName, string destination)
     {
-        var imagePath = Path.Combine(_webHostEnvironment.ContentRootPath, @$"Resources/{destination}", imageName);
-        if (System.IO.File.Exists(imagePath))
-            System.IO.File.Delete(imagePath);
+        if (!string.IsNullOrEmpty(imageName))
+        {
+            var imagePath = Path.Combine(_webHostEnvironment.ContentRootPath, @$"Resources/{destination}", imageName);
+            if (System.IO.File.Exists(imagePath))
+                System.IO.File.Delete(imagePath);
+        }
     }
 
     public async Task<string> SaveImage(IFormFile imageFile, string destination)
